Provide a clean and easy way to make requests for push message delivery #6
With this commit, we introduce components which make it easier for
applications to send requests for the delivery of push messages(we call
these components 'Preparer'). These components have a specific
implementation for each third-party library. Therefore, they have a
specific dependency on the third-party library. We support OkHttp,
Apache HttpClient, Jetty Client and Vert.x Web Client. In order not to
force users to have dependencies on them, these dependencies are
'optional'.

In this commit, we also sightly improved existing javadocs and APIs.

This commit deprecates VAPIDJWTParam.Builder#expiresAfterSeconds(int)
in order to prefer VAPIDJWTParam.Builder#expiresAfter(int, TimeUnit).

Resolves #6.